Move `ProductImageStatus` to its own file.
