+++ WooCommerce/Classes/ViewRelated/Products/Media/ProductImageStatus.swift
@@ -1,0 +1,29 @@
+import Photos
+import Yosemite
+
+/// The status of a Product image.
+///
+enum ProductImageStatus {
+    /// A `PHAsset` is being uploaded.
+    ///
+    case uploading(asset: PHAsset)
+
+    /// The Product image exists remotely.
+    ///
+    case remote(image: ProductImage)
+}
+
+extension ProductImageStatus {
+    var cellReuseIdentifier: String {
+        return cellClass.reuseIdentifier
+    }
+
+    private var cellClass: UICollectionViewCell.Type {
+        switch self {
+        case .uploading:
+            return InProgressProductImageCollectionViewCell.self
+        case .remote:
+            return ProductImageCollectionViewCell.self
+        }
+    }
+}
