+++ WooCommerce/Classes/ViewRelated/Products/Media/ProductImagesCollectionViewController.swift
@@ -2,33 +2,6 @@ import Photos
 import UIKit
 import Yosemite
 
-/// The status of a Product image.
-///
-enum ProductImageStatus {
-    /// A `PHAsset` is being uploaded.
-    ///
-    case uploading(asset: PHAsset)
-
-    /// The Product image exists remotely.
-    ///
-    case remote(image: ProductImage)
-}
-
-extension ProductImageStatus {
-    var cellReuseIdentifier: String {
-        return cellClass.reuseIdentifier
-    }
-
-    private var cellClass: UICollectionViewCell.Type {
-        switch self {
-        case .uploading:
-            return InProgressProductImageCollectionViewCell.self
-        case .remote:
-            return ProductImageCollectionViewCell.self
-        }
-    }
-}
-
 /// Displays Product images in grid layout.
 ///
 final class ProductImagesCollectionViewController: UICollectionViewController {
